requirements.txt
Flask==3.0.3
pandas==2.2.2
scikit-learn==1.4.2

google-generativeai==0.5.4
python-dotenv==1.0.1
